feat: ✨Setup Routes
