+++ src/App.jsx
@@ -1,6 +1,6 @@
 import SignIn from './Routes/SignIn'
 import SignUp from './Routes/SignUp'
-import Home from './Routes/Home'
+
 import { createBrowserRouter } from 'react-router-dom'
 import { RouterProvider } from 'react-router-dom'
 import AuthContext from './Context/AuthContext'
@@ -8,8 +8,8 @@ import Protected from './Routes/Protected'
 import Authentification from './pages/authentification'
 import DaeRender from './components/DaeRender'
 import HomePage from './pages/homePage'
-import SearchDAE from './pages/SearchDAE'
 import InfoDAE from './pages/infoDAE'
+import Profil from './pages/profil'
 
 // TODO : Ici mettre les routes de notre projet
 
@@ -17,42 +17,39 @@ function App() {
   const router = createBrowserRouter([
     {
       path: '/',
-      element: (
-        <Protected>
-          {/* <Home /> */}
-          <DaeRender />
-        </Protected>
-      ),
-    },
-    {
-      path: '/homePage',
       element: <HomePage />,
     },
     {
-      path: '/login',
+      path: '/searchDAE',
+      element: <DaeRender />,
+    },
+    {
+      path: '/info',
+      element: <InfoDAE />,
+    },
+
+    {
+      path: '/auth',
       element: <Authentification />,
     },
     {
-      path: '/sign-in',
-      element: <SignIn />,
+      path: '/profil',
+      element: (
+        <Protected>
+          <Profil />
+        </Protected>
+      ),
     },
     {
       path: '/sign-up',
       element: <SignUp />,
     },
     {
-      path: '/recherche',
-      element: <SearchDAE />,
-    },
-    {
-      path: '/info',
-      element: <InfoDAE />,
+      path: '/sign-in',
+      element: <SignIn />,
     },
   ])
   return (
-    // <>
-    //   <DaeRender />
-    // </>
     <AuthContext>
       <RouterProvider router={router} />
     </AuthContext>
