+++ src/Routes/Home.jsx
@@ -1,22 +1,22 @@
-import { Button } from '@/components/ui/button'
-import { getAuth, signOut } from 'firebase/auth'
+// import { Button } from '@/components/ui/button'
+// import { getAuth, signOut } from 'firebase/auth'
 
-const Home = () => {
-  const auth = getAuth()
-  const handleSignOut = async () => {
-    try {
-      await signOut(auth)
-    } catch (error) {
-      console.log(error)
-    }
-  }
+// const Home = () => {
+//   const auth = getAuth()
+//   const handleSignOut = async () => {
+//     try {
+//       await signOut(auth)
+//     } catch (error) {
+//       console.log(error)
+//     }
+//   }
 
-  return (
-    <div>
-      <h1>This is the home page</h1>
-      <Button onClick={() => handleSignOut()}>Log out</Button>
-    </div>
-  )
-}
+//   return (
+//     <div>
+//       <h1>This is the home page</h1>
+//       <Button onClick={() => handleSignOut()}>Log out</Button>
+//     </div>
+//   )
+// }
 
-export default Home
+// export default Home
