+++ src/Routes/Protected.jsx
@@ -5,7 +5,7 @@ import { Navigate } from 'react-router-dom'
 const Protected = ({ children }) => {
   const { user } = useContext(Context)
   if (!user) {
-    return <Navigate to="/authentification" replace />
+    return <Navigate to="/" replace />
   } else {
     return children
   }
