+++ src/components/DaeRender.jsx
@@ -1,7 +1,5 @@
-import { getAuth, signOut } from 'firebase/auth'
 import { Loader2 } from 'lucide-react'
 import useFetchDAE from '../hooks/useFetchDAE'
-import { Button } from './ui/button'
 
 //Mini component to display loader
 const Loader = () => {
@@ -28,9 +26,6 @@ const DAEList = ({ data }) => {
     <>
       <p className="bg-secondary text-4xl items-center w-full">Liste des DAE</p>
 
-      {/* <div className="mb-6 text-center text-3xl underline decoration-pink-500 ">
-        Nombre de DAE : {data.length}
-      </div> */}
       <div className="grid grid-cols-2">
         {data?.map((entry) => (
           <div key={entry.gid} className="m-6">
@@ -49,20 +44,10 @@ const DAEList = ({ data }) => {
   )
 }
 const DaeRender = () => {
-  const auth = getAuth()
-  const handleSignOut = async () => {
-    try {
-      await signOut(auth)
-    } catch (error) {
-      console.log(error)
-    }
-  }
   const { data, isLoading, error } = useFetchDAE()
-  console.log('test')
-  console.log(data)
+
   return (
     <>
-      <Button onClick={() => handleSignOut()}>Log out</Button>
       {isLoading && <Loader />}
       {!isLoading && !error && <DAEList data={data} />}
       {error && <ErrorMessage message={error} />}
